2008-03-06 Chris Johns <cjohns@users.sourceforge.net>
        * driver/bdm.c: Move the CPU32 and CF specific parts into separate
        files and add TBLCF. Make all call happen via handlers. Move the
        common bit bashing routines into common functions for use by the
        CPU32 and CF driver parts. Add the swap code taken from the
        library code.

        * driver/bdm.h: Change BDMioctl fields to unsigned long int.

        * driver/bdm-cf-pe.c: Moved from bdm.c. Fixed the system reg read
        and write routines to detect a non-mapped register.

        * driver/bdm-cpu32: Moved from bdm.c. Fixed the system reg read
        and write routines to detect a non-mapped register.

        * driver/bdm-tblcf.c: New.

        * driver/ioperm/ioperm.c, driver/win/win-bdm.c: Add USB support.

        * lib/bdmIO.c: Remove swap code. Moved to the driver as TBLCF
        handles the swap. This is the simplest way to untangle this.

        * gdbserver/Makefile.am, lib/Makefile.am, server/Makefile.am,
        test/Makefile.am, utils/Makefile.am: Add TBLCF and libusb support.

        * tblcf/.cvsignore, tblcf/Makefile.am, tblcf/commands.h,
        tblcf/log.c, tblcf/log.h, tblcf/log_cmdline.c,
        tblcf/log_cmdline.h, tblcf/srec.c, tblcf/srec.h,
        tblcf/tblcf-show.c, tblcf/tblcf.c, tblcf/tblcf.h,
        tblcf/tblcf_bt.c, tblcf/tblcf_bt.h, tblcf/tblcf_hwdesc.h,
        tblcf/tblcf_unsec.c, tblcf/tblcf_usb.c, tblcf/tblcf_usb.h,
        tblcf/version.h: New. Taken from the TBLCF project.

        * configure.ac, Makefile.am: Add TBLCF and libusb support.

+++ m68k/driver/bdm.h
@@ -40,11 +40,7 @@
  *
  * Coldfire support by:
  * Chris Johns
- * Objective Design Systems
- * 35 Cairo Street
- * Cammeray, Sydney, 2062, Australia
- *
- * ccj@acm.org
+ * chris@contemporary.net.au
  *
  */
 
@@ -109,8 +105,8 @@
  * Address and value are in host-endian order
  */
 struct BDMioctl {
-    unsigned int    address;
-    unsigned int    value;
+    unsigned long int address;
+    unsigned long int value;
 };
 
 /*
